Added basic support for custom date ranges
A modal containing the new graph can be opened from the graph detail page. For, now this is just the "day" graph without
controls to alter the start/end timestamp (coming later).
The graph data endpoint accepts 'start' and 'end' as GET parameter to specify the date range we're interested in. Will only work with scope=custom.